big update

+++ app/src/main/java/by/bsu/slabko/vladislav/pharmhelper/activities/searchResult/objects/ItemSlot.java
@@ -2,10 +2,10 @@ package by.bsu.slabko.vladislav.pharmhelper.activities.searchResult.objects;
 
 public class ItemSlot {
     public String name;
-    public float price;
+    public double price;
     public int quantity;
     public String manuf;
-    public ItemSlot(String pharmName, float price, int quantity, String manuf){
+    public ItemSlot(String pharmName, double price, int quantity, String manuf){
         this.name = pharmName;
         this.price = price;
         this.quantity = quantity;
